refactor: rename label sync service import and consolidate integration tests
- Updated import path for `syncLabelsFromFile` in `zip-import.service.ts`.
- Added comprehensive integration tests for label synchronization in `labels.service.sync.integration.test.ts`, consolidating tests from `label-sync.service.ts` and `gitlab-file-sync.service.ts`.
- Included various test cases for validation, syncing, and error handling related to label synchronization.
