+++ apps/backend/src/routes/gitlab.routes.ts
@@ -48,7 +48,7 @@ import {
   isValidConflictResolution,
 } from "../lib/validation.js";
 import { createProject, deleteProject } from "../services/projects.service.js";
-import { syncLabelsFromGitLabFile } from "../services/gitlab-file-sync.service.js";
+import { syncLabelsFromGitLabFile } from "../services/labels.service.js";
 
 /**
  * Helper to get the authenticated user ID from a request.
